+++ modules/test/persistence-test/src/testIntegration/java/com/liferay/portal/service/persistence/test/CompanyInfoPersistenceTest.java
@@ -113,7 +113,33 @@ public class CompanyInfoPersistenceTest {
 
 		newCompanyInfo.setCompanyId(RandomTestUtil.nextLong());
 
+		newCompanyInfo.setHomeURL(RandomTestUtil.randomString());
+
+		newCompanyInfo.setIndexNameCurrent(RandomTestUtil.randomString());
+
+		newCompanyInfo.setIndexNameNext(RandomTestUtil.randomString());
+
+		newCompanyInfo.setIndustry(RandomTestUtil.randomString());
+
 		newCompanyInfo.setKey(RandomTestUtil.randomString());
+
+		newCompanyInfo.setLegalId(RandomTestUtil.randomString());
+
+		newCompanyInfo.setLegalName(RandomTestUtil.randomString());
+
+		newCompanyInfo.setLegalType(RandomTestUtil.randomString());
+
+		newCompanyInfo.setLogoId(RandomTestUtil.nextLong());
+
+		newCompanyInfo.setName(RandomTestUtil.randomString());
+
+		newCompanyInfo.setSicCode(RandomTestUtil.randomString());
+
+		newCompanyInfo.setSize(RandomTestUtil.randomString());
+
+		newCompanyInfo.setTickerSymbol(RandomTestUtil.randomString());
+
+		newCompanyInfo.setType(RandomTestUtil.randomString());
 
 		newCompanyInfo = _persistence.update(newCompanyInfo);
 
@@ -131,7 +157,36 @@ public class CompanyInfoPersistenceTest {
 		Assert.assertEquals(
 			existingCompanyInfo.getCompanyId(), newCompanyInfo.getCompanyId());
 		Assert.assertEquals(
+			existingCompanyInfo.getHomeURL(), newCompanyInfo.getHomeURL());
+		Assert.assertEquals(
+			existingCompanyInfo.getIndexNameCurrent(),
+			newCompanyInfo.getIndexNameCurrent());
+		Assert.assertEquals(
+			existingCompanyInfo.getIndexNameNext(),
+			newCompanyInfo.getIndexNameNext());
+		Assert.assertEquals(
+			existingCompanyInfo.getIndustry(), newCompanyInfo.getIndustry());
+		Assert.assertEquals(
 			existingCompanyInfo.getKey(), newCompanyInfo.getKey());
+		Assert.assertEquals(
+			existingCompanyInfo.getLegalId(), newCompanyInfo.getLegalId());
+		Assert.assertEquals(
+			existingCompanyInfo.getLegalName(), newCompanyInfo.getLegalName());
+		Assert.assertEquals(
+			existingCompanyInfo.getLegalType(), newCompanyInfo.getLegalType());
+		Assert.assertEquals(
+			existingCompanyInfo.getLogoId(), newCompanyInfo.getLogoId());
+		Assert.assertEquals(
+			existingCompanyInfo.getName(), newCompanyInfo.getName());
+		Assert.assertEquals(
+			existingCompanyInfo.getSicCode(), newCompanyInfo.getSicCode());
+		Assert.assertEquals(
+			existingCompanyInfo.getSize(), newCompanyInfo.getSize());
+		Assert.assertEquals(
+			existingCompanyInfo.getTickerSymbol(),
+			newCompanyInfo.getTickerSymbol());
+		Assert.assertEquals(
+			existingCompanyInfo.getType(), newCompanyInfo.getType());
 	}
 
 	@Test
@@ -167,7 +222,10 @@ public class CompanyInfoPersistenceTest {
 	protected OrderByComparator<CompanyInfo> getOrderByComparator() {
 		return OrderByComparatorFactoryUtil.create(
 			"CompanyInfo", "mvccVersion", true, "companyInfoId", true,
-			"companyId", true);
+			"companyId", true, "homeURL", true, "indexNameCurrent", true,
+			"indexNameNext", true, "industry", true, "legalId", true,
+			"legalName", true, "legalType", true, "logoId", true, "name", true,
+			"sicCode", true, "size", true, "tickerSymbol", true, "type", true);
 	}
 
 	@Test
@@ -444,7 +502,33 @@ public class CompanyInfoPersistenceTest {
 
 		companyInfo.setCompanyId(RandomTestUtil.nextLong());
 
+		companyInfo.setHomeURL(RandomTestUtil.randomString());
+
+		companyInfo.setIndexNameCurrent(RandomTestUtil.randomString());
+
+		companyInfo.setIndexNameNext(RandomTestUtil.randomString());
+
+		companyInfo.setIndustry(RandomTestUtil.randomString());
+
 		companyInfo.setKey(RandomTestUtil.randomString());
+
+		companyInfo.setLegalId(RandomTestUtil.randomString());
+
+		companyInfo.setLegalName(RandomTestUtil.randomString());
+
+		companyInfo.setLegalType(RandomTestUtil.randomString());
+
+		companyInfo.setLogoId(RandomTestUtil.nextLong());
+
+		companyInfo.setName(RandomTestUtil.randomString());
+
+		companyInfo.setSicCode(RandomTestUtil.randomString());
+
+		companyInfo.setSize(RandomTestUtil.randomString());
+
+		companyInfo.setTickerSymbol(RandomTestUtil.randomString());
+
+		companyInfo.setType(RandomTestUtil.randomString());
 
 		_companyInfos.add(_persistence.update(companyInfo));
 
@@ -456,4 +540,4 @@ public class CompanyInfoPersistenceTest {
 	private ClassLoader _dynamicQueryClassLoader;
 
 }
-// LIFERAY-SERVICE-BUILDER-HASH:1086181586
+// LIFERAY-SERVICE-BUILDER-HASH:313530594
