+++ portal-impl/src/com/liferay/portal/service/persistence/impl/CompanyPersistenceImpl.java
@@ -9,7 +9,6 @@ import com.liferay.portal.kernel.dao.orm.EntityCache;
 import com.liferay.portal.kernel.dao.orm.EntityCacheUtil;
 import com.liferay.portal.kernel.dao.orm.FinderCache;
 import com.liferay.portal.kernel.dao.orm.FinderCacheUtil;
-import com.liferay.portal.kernel.dao.orm.FinderPath;
 import com.liferay.portal.kernel.dao.orm.Session;
 import com.liferay.portal.kernel.exception.NoSuchCompanyException;
 import com.liferay.portal.kernel.model.Company;
@@ -19,10 +18,8 @@ import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.service.persistence.CompanyPersistence;
 import com.liferay.portal.kernel.service.persistence.CompanyUtil;
 import com.liferay.portal.kernel.service.persistence.impl.BasePersistenceImpl;
-import com.liferay.portal.kernel.service.persistence.impl.CollectionPersistenceFinder;
 import com.liferay.portal.kernel.service.persistence.impl.FinderColumn;
 import com.liferay.portal.kernel.service.persistence.impl.UniquePersistenceFinder;
-import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.util.ProxyUtil;
 import com.liferay.portal.kernel.util.SetUtil;
 import com.liferay.portal.model.impl.CompanyImpl;
@@ -34,7 +31,6 @@ import java.lang.reflect.InvocationHandler;
 
 import java.util.Date;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
@@ -121,96 +117,10 @@ public class CompanyPersistenceImpl
 			FinderCacheUtil.getFinderCache(), new Object[] {webId});
 	}
 
-	private CollectionPersistenceFinder<Company, NoSuchCompanyException>
-		_collectionPersistenceFinderByLogoId;
-
-	/**
-	 * Returns an ordered range of all the companies where logoId = &#63;.
-	 *
-	 * <p>
-	 * Useful when paginating results. Returns a maximum of <code>end - start</code> instances. <code>start</code> and <code>end</code> are not primary keys, they are indexes in the result set. Thus, <code>0</code> refers to the first result in the set. Setting both <code>start</code> and <code>end</code> to <code>com.liferay.portal.kernel.dao.orm.QueryUtil#ALL_POS</code> will return the full result set. If <code>orderByComparator</code> is specified, then the query will include the given ORDER BY logic. If <code>orderByComparator</code> is absent, then the query will include the default ORDER BY logic from <code>CompanyModelImpl</code>.
-	 * </p>
-	 *
-	 * @param logoId the logo ID
-	 * @param start the lower bound of the range of companies
-	 * @param end the upper bound of the range of companies (not inclusive)
-	 * @param orderByComparator the comparator to order the results by (optionally <code>null</code>)
-	 * @param useFinderCache whether to use the finder cache
-	 * @return the ordered range of matching companies
-	 */
-	@Override
-	public List<Company> findByLogoId(
-		long logoId, int start, int end,
-		OrderByComparator<Company> orderByComparator, boolean useFinderCache) {
-
-		return _collectionPersistenceFinderByLogoId.find(
-			FinderCacheUtil.getFinderCache(), new Object[] {logoId}, start, end,
-			orderByComparator, useFinderCache);
-	}
-
-	/**
-	 * Returns the first company in the ordered set where logoId = &#63;.
-	 *
-	 * @param logoId the logo ID
-	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
-	 * @return the first matching company
-	 * @throws NoSuchCompanyException if a matching company could not be found
-	 */
-	@Override
-	public Company findByLogoId_First(
-			long logoId, OrderByComparator<Company> orderByComparator)
-		throws NoSuchCompanyException {
-
-		return _collectionPersistenceFinderByLogoId.findFirst(
-			FinderCacheUtil.getFinderCache(), new Object[] {logoId},
-			orderByComparator);
-	}
-
-	/**
-	 * Returns the first company in the ordered set where logoId = &#63;.
-	 *
-	 * @param logoId the logo ID
-	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
-	 * @return the first matching company, or <code>null</code> if a matching company could not be found
-	 */
-	@Override
-	public Company fetchByLogoId_First(
-		long logoId, OrderByComparator<Company> orderByComparator) {
-
-		return _collectionPersistenceFinderByLogoId.fetchFirst(
-			FinderCacheUtil.getFinderCache(), new Object[] {logoId},
-			orderByComparator);
-	}
-
-	/**
-	 * Removes all the companies where logoId = &#63; from the database.
-	 *
-	 * @param logoId the logo ID
-	 */
-	@Override
-	public void removeByLogoId(long logoId) {
-		_collectionPersistenceFinderByLogoId.remove(
-			FinderCacheUtil.getFinderCache(), new Object[] {logoId});
-	}
-
-	/**
-	 * Returns the number of companies where logoId = &#63;.
-	 *
-	 * @param logoId the logo ID
-	 * @return the number of matching companies
-	 */
-	@Override
-	public int countByLogoId(long logoId) {
-		return _collectionPersistenceFinderByLogoId.count(
-			FinderCacheUtil.getFinderCache(), new Object[] {logoId});
-	}
-
 	public CompanyPersistenceImpl() {
 		Map<String, String> dbColumnNames = new HashMap<String, String>();
 
 		dbColumnNames.put("active", "active_");
-		dbColumnNames.put("type", "type_");
-		dbColumnNames.put("size", "size_");
 
 		setDBColumnNames(dbColumnNames);
 
@@ -420,32 +330,6 @@ public class CompanyPersistenceImpl
 				"company.", "webId", FinderColumn.Type.STRING, "=", true, true,
 				Company::getWebId));
 
-		_collectionPersistenceFinderByLogoId =
-			new CollectionPersistenceFinder<>(
-				this,
-				new FinderPath(
-					FINDER_CLASS_NAME_LIST_WITH_PAGINATION, "findByLogoId",
-					new String[] {
-						Long.class.getName(), Integer.class.getName(),
-						Integer.class.getName(),
-						OrderByComparator.class.getName()
-					},
-					new String[] {"logoId"}, true),
-				new FinderPath(
-					FINDER_CLASS_NAME_LIST_WITHOUT_PAGINATION, "findByLogoId",
-					new String[] {Long.class.getName()},
-					new String[] {"logoId"}, true),
-				new FinderPath(
-					FINDER_CLASS_NAME_LIST_WITHOUT_PAGINATION, "countByLogoId",
-					new String[] {Long.class.getName()},
-					new String[] {"logoId"}, false),
-				_SQL_SELECT_COMPANY_WHERE, _SQL_COUNT_COMPANY_WHERE,
-				CompanyModelImpl.ORDER_BY_JPQL, _ENTITY_ALIAS_PREFIX, "", "",
-				null,
-				new FinderColumn<>(
-					"company.", "logoId", FinderColumn.Type.LONG, "=", true,
-					true, Company::getLogoId));
-
 		CompanyUtil.setPersistence(this);
 	}
 
@@ -455,20 +339,14 @@ public class CompanyPersistenceImpl
 		EntityCacheUtil.removeCache(CompanyImpl.class.getName());
 	}
 
-	private static final String _ENTITY_ALIAS_PREFIX =
-		CompanyModelImpl.ENTITY_ALIAS + ".";
-
 	private static final String _SQL_SELECT_COMPANY =
 		"SELECT company FROM Company company";
 
 	private static final String _SQL_SELECT_COMPANY_WHERE =
 		"SELECT company FROM Company company WHERE ";
 
-	private static final String _SQL_COUNT_COMPANY_WHERE =
-		"SELECT COUNT(company) FROM Company company WHERE ";
-
 	private static final Set<String> _badColumnNames = SetUtil.fromArray(
-		new String[] {"active", "type", "size"});
+		new String[] {"active"});
 
 	@Override
 	protected FinderCache getFinderCache() {
@@ -476,4 +354,4 @@ public class CompanyPersistenceImpl
 	}
 
 }
-// LIFERAY-SERVICE-BUILDER-HASH:169398529
+// LIFERAY-SERVICE-BUILDER-HASH:-1121413014
