+++ portal-kernel/src/com/liferay/portal/kernel/model/CompanyModel.java
@@ -190,35 +190,6 @@ public interface CompanyModel
 	public void setMx(String mx);
 
 	/**
-	 * Returns the home url of this company.
-	 *
-	 * @return the home url of this company
-	 */
-	@AutoEscape
-	public String getHomeURL();
-
-	/**
-	 * Sets the home url of this company.
-	 *
-	 * @param homeURL the home url of this company
-	 */
-	public void setHomeURL(String homeURL);
-
-	/**
-	 * Returns the logo ID of this company.
-	 *
-	 * @return the logo ID of this company
-	 */
-	public long getLogoId();
-
-	/**
-	 * Sets the logo ID of this company.
-	 *
-	 * @param logoId the logo ID of this company
-	 */
-	public void setLogoId(long logoId);
-
-	/**
 	 * Returns the max users of this company.
 	 *
 	 * @return the max users of this company
@@ -253,171 +224,6 @@ public interface CompanyModel
 	 */
 	public void setActive(boolean active);
 
-	/**
-	 * Returns the name of this company.
-	 *
-	 * @return the name of this company
-	 */
-	@AutoEscape
-	public String getName();
-
-	/**
-	 * Sets the name of this company.
-	 *
-	 * @param name the name of this company
-	 */
-	public void setName(String name);
-
-	/**
-	 * Returns the legal name of this company.
-	 *
-	 * @return the legal name of this company
-	 */
-	@AutoEscape
-	public String getLegalName();
-
-	/**
-	 * Sets the legal name of this company.
-	 *
-	 * @param legalName the legal name of this company
-	 */
-	public void setLegalName(String legalName);
-
-	/**
-	 * Returns the legal ID of this company.
-	 *
-	 * @return the legal ID of this company
-	 */
-	@AutoEscape
-	public String getLegalId();
-
-	/**
-	 * Sets the legal ID of this company.
-	 *
-	 * @param legalId the legal ID of this company
-	 */
-	public void setLegalId(String legalId);
-
-	/**
-	 * Returns the legal type of this company.
-	 *
-	 * @return the legal type of this company
-	 */
-	@AutoEscape
-	public String getLegalType();
-
-	/**
-	 * Sets the legal type of this company.
-	 *
-	 * @param legalType the legal type of this company
-	 */
-	public void setLegalType(String legalType);
-
-	/**
-	 * Returns the sic code of this company.
-	 *
-	 * @return the sic code of this company
-	 */
-	@AutoEscape
-	public String getSicCode();
-
-	/**
-	 * Sets the sic code of this company.
-	 *
-	 * @param sicCode the sic code of this company
-	 */
-	public void setSicCode(String sicCode);
-
-	/**
-	 * Returns the ticker symbol of this company.
-	 *
-	 * @return the ticker symbol of this company
-	 */
-	@AutoEscape
-	public String getTickerSymbol();
-
-	/**
-	 * Sets the ticker symbol of this company.
-	 *
-	 * @param tickerSymbol the ticker symbol of this company
-	 */
-	public void setTickerSymbol(String tickerSymbol);
-
-	/**
-	 * Returns the industry of this company.
-	 *
-	 * @return the industry of this company
-	 */
-	@AutoEscape
-	public String getIndustry();
-
-	/**
-	 * Sets the industry of this company.
-	 *
-	 * @param industry the industry of this company
-	 */
-	public void setIndustry(String industry);
-
-	/**
-	 * Returns the type of this company.
-	 *
-	 * @return the type of this company
-	 */
-	@AutoEscape
-	public String getType();
-
-	/**
-	 * Sets the type of this company.
-	 *
-	 * @param type the type of this company
-	 */
-	public void setType(String type);
-
-	/**
-	 * Returns the size of this company.
-	 *
-	 * @return the size of this company
-	 */
-	@AutoEscape
-	public String getSize();
-
-	/**
-	 * Sets the size of this company.
-	 *
-	 * @param size the size of this company
-	 */
-	public void setSize(String size);
-
-	/**
-	 * Returns the index name current of this company.
-	 *
-	 * @return the index name current of this company
-	 */
-	@AutoEscape
-	public String getIndexNameCurrent();
-
-	/**
-	 * Sets the index name current of this company.
-	 *
-	 * @param indexNameCurrent the index name current of this company
-	 */
-	public void setIndexNameCurrent(String indexNameCurrent);
-
-	/**
-	 * Returns the index name next of this company.
-	 *
-	 * @return the index name next of this company
-	 */
-	@AutoEscape
-	public String getIndexNameNext();
-
-	/**
-	 * Sets the index name next of this company.
-	 *
-	 * @param indexNameNext the index name next of this company
-	 */
-	public void setIndexNameNext(String indexNameNext);
-
 	@Override
 	public Company cloneWithOriginalValues();
 
@@ -426,4 +232,4 @@ public interface CompanyModel
 	}
 
 }
-// LIFERAY-SERVICE-BUILDER-HASH:-269727918
+// LIFERAY-SERVICE-BUILDER-HASH:569423917
