+++ portal-kernel/src/com/liferay/portal/kernel/service/persistence/CompanyPersistence.java
@@ -65,67 +65,6 @@ public interface CompanyPersistence extends BasePersistence<Company> {
 	public int countByWebId(String webId);
 
 	/**
-	 * Returns an ordered range of all the companies where logoId = &#63;.
-	 *
-	 * <p>
-	 * Useful when paginating results. Returns a maximum of <code>end - start</code> instances. <code>start</code> and <code>end</code> are not primary keys, they are indexes in the result set. Thus, <code>0</code> refers to the first result in the set. Setting both <code>start</code> and <code>end</code> to <code>com.liferay.portal.kernel.dao.orm.QueryUtil#ALL_POS</code> will return the full result set. If <code>orderByComparator</code> is specified, then the query will include the given ORDER BY logic. If <code>orderByComparator</code> is absent, then the query will include the default ORDER BY logic from <code>com.liferay.portal.model.impl.CompanyModelImpl</code>.
-	 * </p>
-	 *
-	 * @param logoId the logo ID
-	 * @param start the lower bound of the range of companies
-	 * @param end the upper bound of the range of companies (not inclusive)
-	 * @param orderByComparator the comparator to order the results by (optionally <code>null</code>)
-	 * @param useFinderCache whether to use the finder cache
-	 * @return the ordered range of matching companies
-	 */
-	public java.util.List<Company> findByLogoId(
-		long logoId, int start, int end,
-		com.liferay.portal.kernel.util.OrderByComparator<Company>
-			orderByComparator,
-		boolean useFinderCache);
-
-	/**
-	 * Returns the first company in the ordered set where logoId = &#63;.
-	 *
-	 * @param logoId the logo ID
-	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
-	 * @return the first matching company
-	 * @throws NoSuchCompanyException if a matching company could not be found
-	 */
-	public Company findByLogoId_First(
-			long logoId,
-			com.liferay.portal.kernel.util.OrderByComparator<Company>
-				orderByComparator)
-		throws NoSuchCompanyException;
-
-	/**
-	 * Returns the first company in the ordered set where logoId = &#63;.
-	 *
-	 * @param logoId the logo ID
-	 * @param orderByComparator the comparator to order the set by (optionally <code>null</code>)
-	 * @return the first matching company, or <code>null</code> if a matching company could not be found
-	 */
-	public Company fetchByLogoId_First(
-		long logoId,
-		com.liferay.portal.kernel.util.OrderByComparator<Company>
-			orderByComparator);
-
-	/**
-	 * Removes all the companies where logoId = &#63; from the database.
-	 *
-	 * @param logoId the logo ID
-	 */
-	public void removeByLogoId(long logoId);
-
-	/**
-	 * Returns the number of companies where logoId = &#63;.
-	 *
-	 * @param logoId the logo ID
-	 * @return the number of matching companies
-	 */
-	public int countByLogoId(long logoId);
-
-	/**
 	 * Creates a new company with the primary key. Does not add the company to the database.
 	 *
 	 * @param companyId the primary key for the new company
@@ -172,56 +111,5 @@ public interface CompanyPersistence extends BasePersistence<Company> {
 		return fetchByWebId(webId, true);
 	}
 
-	/**
-	 * Returns all the companies where logoId = &#63;.
-	 *
-	 * @param logoId the logo ID
-	 * @return the matching companies
-	 */
-	public default java.util.List<Company> findByLogoId(long logoId) {
-		return findByLogoId(
-			logoId, com.liferay.portal.kernel.dao.orm.QueryUtil.ALL_POS,
-			com.liferay.portal.kernel.dao.orm.QueryUtil.ALL_POS, null, true);
-	}
-
-	/**
-	 * Returns a range of all the companies where logoId = &#63;.
-	 *
-	 * <p>
-	 * Useful when paginating results. Returns a maximum of <code>end - start</code> instances. <code>start</code> and <code>end</code> are not primary keys, they are indexes in the result set. Thus, <code>0</code> refers to the first result in the set. Setting both <code>start</code> and <code>end</code> to <code>com.liferay.portal.kernel.dao.orm.QueryUtil#ALL_POS</code> will return the full result set. If <code>orderByComparator</code> is specified, then the query will include the given ORDER BY logic. If <code>orderByComparator</code> is absent, then the query will include the default ORDER BY logic from <code>com.liferay.portal.model.impl.CompanyModelImpl</code>.
-	 * </p>
-	 *
-	 * @param logoId the logo ID
-	 * @param start the lower bound of the range of companies
-	 * @param end the upper bound of the range of companies (not inclusive)
-	 * @return the range of matching companies
-	 */
-	public default java.util.List<Company> findByLogoId(
-		long logoId, int start, int end) {
-
-		return findByLogoId(logoId, start, end, null, true);
-	}
-
-	/**
-	 * Returns an ordered range of all the companies where logoId = &#63;.
-	 *
-	 * <p>
-	 * Useful when paginating results. Returns a maximum of <code>end - start</code> instances. <code>start</code> and <code>end</code> are not primary keys, they are indexes in the result set. Thus, <code>0</code> refers to the first result in the set. Setting both <code>start</code> and <code>end</code> to <code>com.liferay.portal.kernel.dao.orm.QueryUtil#ALL_POS</code> will return the full result set. If <code>orderByComparator</code> is specified, then the query will include the given ORDER BY logic. If <code>orderByComparator</code> is absent, then the query will include the default ORDER BY logic from <code>com.liferay.portal.model.impl.CompanyModelImpl</code>.
-	 * </p>
-	 *
-	 * @param logoId the logo ID
-	 * @param start the lower bound of the range of companies
-	 * @param end the upper bound of the range of companies (not inclusive)
-	 * @param orderByComparator the comparator to order the results by (optionally <code>null</code>)
-	 * @return the ordered range of matching companies
-	 */
-	public default java.util.List<Company> findByLogoId(
-		long logoId, int start, int end,
-		com.liferay.portal.kernel.util.OrderByComparator<Company>
-			orderByComparator) {
-
-		return findByLogoId(logoId, start, end, orderByComparator, true);
-	}
-
 }
-// LIFERAY-SERVICE-BUILDER-HASH:-1744450588
+// LIFERAY-SERVICE-BUILDER-HASH:-1773781155
